requirements.txt
flask
flask_sqlalchemy
flask_cors
flask_jwt_extended
flask_bcrypt
python-dotenv
authlib
requests
psycopg2-binary
pytest-flask